Some cleanup and test enhancements.
