+++ tests/Agent/IntegrationTests/IntegrationTestHelpers/RemoteServiceFixtures/RemoteApplicationFixture.cs
@@ -4,13 +4,11 @@
 
 using System;
 using System.Collections.Generic;
-using System.Diagnostics;
 using System.IO;
 using System.Linq;
 using System.Net;
 using System.Net.Http;
 using System.Threading;
-using System.Threading.Tasks;
 using NewRelic.Agent.IntegrationTests.Shared;
 using Newtonsoft.Json;
 using Xunit;
@@ -335,6 +333,11 @@ namespace NewRelic.Agent.IntegrationTestHelpers.RemoteServiceFixtures
                                     Thread.Sleep(1000);
                                     numberOfTries++;
                                 }
+
+                                if (!applicationHadNonZeroExitCode)
+                                {
+                                    TestForKnownProblems();
+                                }
                             });
                             TestLogger?.WriteLine($"Remote application shutdown time: {timer.Total:N4} seconds");
                         }
@@ -584,66 +587,13 @@ namespace NewRelic.Agent.IntegrationTestHelpers.RemoteServiceFixtures
             Assert.True(result.IsSuccessStatusCode);
         }
 
-    }
-
-
-    // borrowed from XUnit.Sdk.ExecutionTimer, as using their implementation caused a runtime error looking for xunit.execution.dotnet.dll
-    /// <summary>
-    /// Measures and aggregates execution time of one or more actions.
-    /// </summary>
-    public class ExecutionTimer
-    {
-        TimeSpan total;
-
-        /// <summary>
-        /// Returns the total time aggregated across all the actions.
-        /// </summary>
-        public decimal Total
+        // Tests for things like transaction garbage collected, response time already captured, etc.
+        // Works best when logging is at FINEST.
+        private void TestForKnownProblems()
         {
-            get { return (decimal)total.TotalSeconds; }
-        }
-
-        /// <summary>
-        /// Executes an action and aggregates its run time into the total.
-        /// </summary>
-        /// <param name="action">The action to measure.</param>
-        public void Aggregate(Action action)
-        {
-            var stopwatch = Stopwatch.StartNew();
-            try
-            {
-                action();
-            }
-            finally
-            {
-                total += stopwatch.Elapsed;
-            }
-        }
-
-        /// <summary>
-        /// Executes an asynchronous action and aggregates its run time into the total.
-        /// </summary>
-        /// <param name="asyncAction">The action to measure.</param>
-        public async Task AggregateAsync(Func<Task> asyncAction)
-        {
-            var stopwatch = Stopwatch.StartNew();
-            try
-            {
-                await asyncAction();
-            }
-            finally
-            {
-                total += stopwatch.Elapsed;
-            }
-        }
-
-        /// <summary>
-        /// Aggregates a time span into the total time.
-        /// </summary>
-        /// <param name="time">The time to add.</param>
-        public void Aggregate(TimeSpan time)
-        {
-            total += time;
+            Assert.Null(AgentLog.TryGetLogLine(AgentLogBase.TransactionEndedByGCFinalizerLogLineRegEx));
+            Assert.Null(AgentLog.TryGetLogLine(AgentLogBase.TransactionHasAlreadyCapturedResponseTimeLogLineRegEx));
+            TestLogger?.WriteLine("Tests for known problems completed.");
         }
     }
 }
